Improved retries + adapt tests
