+++ modules/flowable-engine-common/src/main/java/org/flowable/common/engine/impl/AbstractEngineConfiguration.java
@@ -57,7 +57,7 @@ import org.flowable.common.engine.impl.db.MybatisTypeAliasConfigurator;
 import org.flowable.common.engine.impl.db.MybatisTypeHandlerConfigurator;
 import org.flowable.common.engine.impl.db.SchemaManager;
 import org.flowable.common.engine.impl.event.EventDispatchAction;
-import org.flowable.common.engine.impl.interceptor.CockroachDbRetryInterceptor;
+import org.flowable.common.engine.impl.interceptor.CrDbRetryInterceptor;
 import org.flowable.common.engine.impl.interceptor.Command;
 import org.flowable.common.engine.impl.interceptor.CommandConfig;
 import org.flowable.common.engine.impl.interceptor.CommandContextFactory;
@@ -504,7 +504,7 @@ public abstract class AbstractEngineConfiguration {
 
             // This probably won't work in environments such as Spring if the transaction is externally managed
             if ("cockroachdb".equals(databaseType)) {
-                interceptors.add(new CockroachDbRetryInterceptor());
+                interceptors.add(new CrDbRetryInterceptor());
             }
 
             CommandInterceptor transactionInterceptor = createTransactionInterceptor();
