+++ modules/flowable-engine/src/test/java/org/flowable/engine/test/api/v6/Flowable6Test.java
@@ -27,7 +27,6 @@ import org.flowable.engine.runtime.Execution;
 import org.flowable.engine.runtime.ProcessInstance;
 import org.flowable.job.api.Job;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 /**
  * These are the first tests ever written for Flowable 6. Keeping them here for nostalgic reasons.
@@ -121,7 +120,6 @@ public class Flowable6Test extends PluggableFlowableTestCase {
      */
     @Test
     @org.flowable.engine.test.Deployment
-    @EnabledIfSystemProperty(named = "db", matches = "crdb")
     public void testLongServiceTaskLoop() {
         int maxCount = 3210; // You can make this as big as you want (as long as it still fits within transaction timeouts).
                              // Go on, try it!
