+++ modules/flowable-engine/src/test/java/org/flowable/engine/test/cfg/taskcount/ChangeTaskCountConfigAndRebootEngineTest.java
@@ -24,9 +24,11 @@ import org.flowable.engine.runtime.ProcessInstance;
 import org.flowable.engine.test.Deployment;
 import org.flowable.task.service.impl.persistence.CountingTaskEntity;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+@DisabledIfSystemProperty(named = "db", matches = "crdb") // Disabled due to having a retry interceptor for CRDB and barriers in this test
 public class ChangeTaskCountConfigAndRebootEngineTest extends ResourceFlowableTestCase {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ChangeTaskCountConfigAndRebootEngineTest.class);
