+++ modules/flowable-engine/src/test/java/org/flowable/engine/test/db/DuplicateVariableInsertTest.java
@@ -27,7 +27,9 @@ import org.flowable.engine.impl.cmd.SetTaskVariablesCmd;
 import org.flowable.engine.impl.test.PluggableFlowableTestCase;
 import org.flowable.engine.runtime.ProcessInstance;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
 
+@DisabledIfSystemProperty(named = "db", matches = "crdb") // Disabled due to having a retry interceptor for CRDB and barriers in this test
 public class DuplicateVariableInsertTest extends PluggableFlowableTestCase {
 
     /**
